% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl.R
\name{vl}
\alias{vl}
\title{vl, a Vega-Lite environment}
\usage{
vl()
}
\description{
Provides an interface to building Vega-Lite JSON specifications
}
\section{Usage}{

\preformatted{
v <- vl()
v$data(mtcars)$
point()$
  x("mpg:Q")$
  y("hp")$
plot()
}
}

